backend/requirements.txt
fastapi==0.72.0
uvicorn==0.17.0
python-jose==3.3.0
passlib==1.7.4
pydantic[email]